+++ Source/School.Api/School.Api.csproj
@@ -6,4 +6,10 @@
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\School.ApplicationCore\School.ApplicationCore.csproj" />
+    <ProjectReference Include="..\School.Business\School.Business.csproj" />
+    <ProjectReference Include="..\School.Data\School.Data.csproj" />
+  </ItemGroup>
+
 </Project>
