+++ Source/School.Api/appsettings.json
@@ -2,7 +2,8 @@
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
+      "Microsoft.AspNetCore": "Information",
+      "Microsoft": "Information"
     }
   },
   "AllowedHosts": "*"
