+++ Source/School.Business/School.Business.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\School.Data\School.Data.csproj" />
+  </ItemGroup>
+
 </Project>
